+++ resources/assets/scripts/ichava-ts/utils/index.ts
@@ -5,4 +5,5 @@
 export { Clipboard, copyToClipboard } from './Clipboard'
 export { Download, downloadSvg } from './Download'
 export { Formatter, formatNumber, formatCompact, formatTimeAgo, formatBytes } from './Formatter'
+export { sanitizeSvg, pickSanitizedSvg } from './sanitizeSvg'